+++ CHANGELOG.md
@@ -235,6 +235,7 @@ You can use any of the following kinds of expressions in a `when:` header:
   - a warning diagnostic will be generated for each empty node
 - Fixed a bug where set-referencing inferred value set statements would crash the compiler
 - The language server no longer truncates XML documentation comments when it reaches a nested XML node. 
+- The constructor for `Yarn.Line` is now public. Previously, it was internal.
 
 ### Removed